Redirect permanently Katuma's alpha to app
Now that we got app.katuma.org up and running we need to tell browsers
and any other HTTP clients that alpha.katuma.org is deprecated, while
not breaking any old link. As long as more browsers cache this, we
should see less traffic hitting the alpha subdomain.

This decouples this temporal solution from ofn-install and makes it
easier to eventually remove it.