Remove Katuma's custom report from production
There's been a single request for this hacky report in the last month
and from the only hub that actively used it (Ridorta). It was someone
from home that did not know how to access the one Ridorta started using
last year.

IMO it's a must we stop serving this custom report now that user
acquisition is growing. If more hubs adopt it, it'll be harder to remove
in the future and we might even need to maintain it and we can't afford
it.

It's been fully confirmed OFN's the built-in packing report provides the
same information as our custom one.